+++ src/components/MyPatientsFolder/MainContents/PatientPrescriptionsSearch.jsx
@@ -1,0 +1,50 @@
+import React from 'react';
+
+import '../../../Style.css'
+
+
+
+
+
+import {Input,Space } from 'antd';
+
+
+
+
+
+
+
+
+
+function PatientPrescriptionsSearch() {
+
+
+ 
+const { Search } = Input;
+const onSearch = (value) => console.log(value);    
+
+
+
+
+
+    return (
+        <>
+         
+  
+          <div style={{margintop:"33px"}}>
+          <Space direction="vertical">
+                <Search
+                    placeholder="Search patients name"
+                    onSearch={onSearch}
+                    className="search-input" />
+            </Space>
+            </div>
+
+         
+     
+         
+        </>
+
+    )
+}
+export default PatientPrescriptionsSearch;
